BE/requirements.txt
alembic==1.13.2
attrs==23.2.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==20.0.4
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
PyJWT==1.7.1
pytest==6.2.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
ruamel.yaml==0.16.5
six==1.16.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
toml==0.10.2
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
